refactor: add three components to split cart page code
