+++ src/components/clear-modal/styles.module.scss
@@ -1,0 +1,70 @@
+.wrapper {
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+  z-index: 1500;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 0 20px;
+  background-color: rgb(0 0 0 / 40%);
+  opacity: 1;
+  visibility: visible;
+  transition: opacity 0.3s, visibility 0.3s;
+  backdrop-filter: grayscale(1);
+
+  &_hidden {
+    opacity: 0;
+    visibility: hidden;
+  }
+
+  & .modal {
+    width: 400px;
+    padding: 30px;
+    background: linear-gradient(var(--background-color) 0 0) padding-box, linear-gradient(45deg, var(--primary-color), var(--secondary-color)) border-box;
+    border: 5px solid transparent;
+    border-radius: 10px;
+
+    & .question {
+      margin-top: 0;
+      font-size: 1.7rem;
+      text-align: center;
+
+      @media (width <= 500px) {
+        font-size: 1.5rem;
+      }
+    }
+
+    & .buttons {
+      display: flex;
+      justify-content: space-around;
+      font-size: 1.5rem;
+
+      @media (width <= 500px) {
+        font-size: 1.2rem;
+      }
+    }
+
+    @media (width <= 500px) {
+      width: 350px;
+      padding: 25px;
+    }
+  }
+}
+
+.confirm {
+  color: var(--text-color);
+  background-image: linear-gradient(45deg, #1ec04c, #17a740);
+
+  &:disabled {
+    background-image: linear-gradient(45deg, #1ec04c, #17a740);
+  }
+}
+
+.cancel {
+  &:disabled {
+    background-image: linear-gradient(45deg, #e43737, var(--error-color));
+  }
+}
