+++ src/components/promocodes/types.ts
@@ -1,0 +1,9 @@
+import type { Cart } from '@commercetools/platform-sdk';
+
+import type { Promocode } from '@/contexts/promocodes-context';
+
+export interface PromocodesProps {
+  addPromocode: (code: string) => Promise<void>;
+  promocodes: Promocode[];
+  cart: Cart;
+}
