+++ src/pages/cart/styles.module.scss
@@ -169,123 +169,6 @@
   }
 }
 
-.total_price_block {
-  flex-shrink: 0;
-  width: 400px;
-  height: max-content;
-  padding: 30px;
-  background: linear-gradient(var(--background-color) 0 0) padding-box, linear-gradient(45deg, var(--primary-color), var(--secondary-color)) border-box;
-  border: 5px solid transparent;
-  border-radius: 10px;
-
-
-  @media (width <= 1300px) {
-    width: 320px;
-    padding: 20px;
-  }
-
-  @media (width <= 950px) {
-    display: flex;
-    flex-direction: column;
-    width: 80%;
-  }
-
-  @media (width <= 450px) {
-    width: 100%;
-  }
-}
-
-.total_price,
-.discount_block,
-.original_price_block {
-  display: flex;
-  justify-content: space-between;
-}
-
-.total_price {
-  margin: 0;
-  font-size: 2rem;
-
-  @media (width <= 1300px) {
-    font-size: 1.4rem;
-  }
-}
-
-.discount_block,
-.original_price_block {
-  margin: 0;
-  font-size: 1.5rem;
-
-  @media (width <= 1300px) {
-    font-size: 1rem;
-  }
-}
-
-.original_price_block {
-  margin-bottom: 0.66em;
-}
-
-.price,
-.discount {
-  font-weight: 600;
-}
-
-.discount {
-  text-decoration: line-through;
-  color: var(--error-color);
-}
-
-.promobox {
-  margin-top: 16px;
-
-  @media (width <= 1300px) {
-    margin-top: 0.66em;
-  }
-}
-
-.promo_desc {
-  margin: 0 0 5px;
-  font-size: 1.3rem;
-
-  @media (width <= 1300px) {
-    margin: 0;
-    font-size: 0.91rem;
-  }
-}
-
-.promoblock {
-  display: flex;
-  gap: 20px;
-  justify-content: space-between;
-  align-items: center;
-  margin: 0 0 24px;
-
-  @media (width <= 1300px) {
-    margin: 0 0 16px;
-  }
-}
-
-.promocode {
-  box-sizing: border-box;
-  width: 100%;
-  height: 30px;
-  padding: 0 5px;
-  font-family: var(--primary-font);
-  font-size: 1em;
-  color: var(--primary-color);
-  border: 1px solid var(--primary-color);
-  border-radius: 6px;
-
-  &:focus {
-    border: 1px solid #000000;
-  }
-
-  &:focus-visible {
-    outline: none;
-  }
-}
-
-.check_out,
 .clear_cart {
   font-size: 1.3rem;
 
@@ -296,97 +179,5 @@
   @media (width <= 950px) {
     width: 240px;
     font-size: 1.3rem;
-  }
-}
-
-.check_out {
-  width: 100%;
-
-  @media (width <= 950px) {
-    width: 240px;
-    margin: auto;
-  }
-}
-
-.promocodes {
-  display: flex;
-  justify-content: space-between;
-  margin: 0;
-  font-size: 1.2rem;
-  color: var(--secondary-color);
-
-  @media (width <= 1300px) {
-    font-size: 0.84rem;
-  }
-}
-
-.wrapper {
-  position: fixed;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  z-index: 1500;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 0 20px;
-  background-color: rgb(0 0 0 / 40%);
-  opacity: 1;
-  visibility: visible;
-  transition: opacity 0.3s, visibility 0.3s;
-  backdrop-filter: grayscale(1);
-
-  &_hidden {
-    opacity: 0;
-    visibility: hidden;
-  }
-
-  & .modal {
-    width: 400px;
-    padding: 30px;
-    background: linear-gradient(var(--background-color) 0 0) padding-box, linear-gradient(45deg, var(--primary-color), var(--secondary-color)) border-box;
-    border: 5px solid transparent;
-    border-radius: 10px;
-
-    & .question {
-      margin-top: 0;
-      font-size: 1.7rem;
-      text-align: center;
-
-      @media (width <= 500px) {
-        font-size: 1.5rem;
-      }
-    }
-
-    & .buttons {
-      display: flex;
-      justify-content: space-around;
-      font-size: 1.5rem;
-
-      @media (width <= 500px) {
-        font-size: 1.2rem;
-      }
-    }
-
-    @media (width <= 500px) {
-      width: 350px;
-      padding: 25px;
-    }
-  }
-}
-
-.confirm {
-  color: var(--text-color);
-  background-image: linear-gradient(45deg, #1ec04c, #17a740);
-
-  &:disabled {
-    background-image: linear-gradient(45deg, #1ec04c, #17a740);
-  }
-}
-
-.cancel {
-  &:disabled {
-    background-image: linear-gradient(45deg, #e43737, var(--error-color));
   }
 }
