+++ src/pages/cart/total-price/styles.module.scss
@@ -1,0 +1,80 @@
+.total_price_block {
+  flex-shrink: 0;
+  width: 400px;
+  height: max-content;
+  padding: 30px;
+  background: linear-gradient(var(--background-color) 0 0) padding-box, linear-gradient(45deg, var(--primary-color), var(--secondary-color)) border-box;
+  border: 5px solid transparent;
+  border-radius: 10px;
+
+
+  @media (width <= 1300px) {
+    width: 320px;
+    padding: 20px;
+  }
+
+  @media (width <= 950px) {
+    display: flex;
+    flex-direction: column;
+    width: 80%;
+  }
+
+  @media (width <= 450px) {
+    width: 100%;
+  }
+}
+
+.total_price,
+.discount_block,
+.original_price_block {
+  display: flex;
+  justify-content: space-between;
+}
+
+.total_price {
+  margin: 0;
+  font-size: 2rem;
+
+  @media (width <= 1300px) {
+    font-size: 1.4rem;
+  }
+}
+
+.price,
+.discount {
+  font-weight: 600;
+}
+
+.discount {
+  text-decoration: line-through;
+  color: var(--error-color);
+}
+
+.discount_block,
+.original_price_block {
+  margin: 0;
+  font-size: 1.5rem;
+
+  @media (width <= 1300px) {
+    font-size: 1rem;
+  }
+}
+
+.original_price_block {
+  margin-bottom: 0.66em;
+}
+
+.check_out {
+  width: 100%;
+  font-size: 1.3rem;
+
+  @media (width <= 1300px) {
+    font-size: 1rem;
+  }
+
+  @media (width <= 950px) {
+    width: 240px;
+    margin: auto;
+    font-size: 1.3rem;
+  }
+}
